services/web/requirements.txt
Flask~=1.1.4
alembic==1.6.5
SQLAlchemy==1.4.20
uWSGI==2.0.19.1
psycopg2-binary==2.8.6
Flask-Migrate==3.0.1
gunicorn==20.1.0
flask_login==0.4.1
flask-restplus==0.13.0
Werkzeug==0.16.1
marshmallow==3.12.2
faker==6.1.1
sqlalchemy==1.4.20
PyYAML==5.4.1
pytest==6.2.4
requests==2.26.0